Value module: extract ArrayOpEngine from evaluate.ts (behavior-preserving)
